current_working_requirements.txt
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
ffmpeg-python==0.2.0
future==1.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.40.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.3.1
oauthlib==3.3.1
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
